hange the code to match Airbnb style
Airbnb's style guide is rather extensive when it comes to such things,
so the majority of the ESLintrc file is shaved away for the closest to
pure "their style" it could be.

The result is rather striking, both in file size and code style.
Personally I am a fan of it, but I can see how some might not be.

Well, that's one candidate down.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>