Apply the Semistandard code style
`eslint-plugin-node` brought up some interesting information:
`assert.equal` is deprecated in Node 10 for `assert.strictEqual`.  So I
went to look, and the latter is [definitely supported][history] by any
version of Node in support today, so there's no reason *not to* change
the code over.  Unless you're writing a fuzzy test for `null` or
`undefined`.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>